Parse individual TimeStampToken
Embedded time stamps do not include the response structure. Renamed "Response" to "Timestamp" to reflect the reflect the more common use case as the
structure is representing the timestamp in the response and not the response itself. Added a Parse function that can parse a timestamp that is not part of a response structure.